#19 refactor: invite 디렉토리 변경으로 인한 수정
