+++ backend/src/main/java/meetween/backend/location/domain/InviteCodeRepository.java
@@ -1,5 +1,6 @@
-package meetween.backend.appointment.domain;
+package meetween.backend.location.domain;
 
+import meetween.backend.invitecode.domain.InviteCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InviteCodeRepository extends JpaRepository<InviteCode, Long> {
