+++ backend/src/test/java/meetween/backend/appointment/application/AppointmentServiceTest.java
@@ -16,6 +16,7 @@ import meetween.backend.appointment.dto.request.AppointmentParticipateRequest;
 import meetween.backend.appointment.dto.response.AppointmentResponse;
 import meetween.backend.appointment.dto.response.IntegratedAppointmentResponses;
 import meetween.backend.category.domain.CategoryRepository;
+import meetween.backend.location.domain.InviteCodeRepository;
 import meetween.backend.location.domain.LocationRepository;
 import meetween.backend.member.domain.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
