+++ backend/src/test/java/meetween/backend/appointment/domain/AppointmentTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
 import meetween.backend.appointment.exception.InvalidAppointmentException;
+import meetween.backend.invitecode.domain.InviteCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
