+++ backend/src/test/java/meetween/backend/support/fixture/common/AppointmentFixtures.java
@@ -1,7 +1,8 @@
 package meetween.backend.support.fixture.common;
 
 import meetween.backend.appointment.domain.Appointment;
-import meetween.backend.appointment.domain.InviteCode;
+import meetween.backend.appointment.dto.response.AppointmentResponse;
+import meetween.backend.invitecode.domain.InviteCode;
 import meetween.backend.appointment.dto.request.AppointmentCreateRequest;
 import meetween.backend.appointment.dto.request.AppointmentParticipateRequest;
 import meetween.backend.category.domain.Category;
@@ -21,6 +22,7 @@ public class AppointmentFixtures {
     public static final BigDecimal 수현_약속_경도 = BigDecimal.valueOf(37.5280674292228);
     public static final AppointmentCreateRequest 수현_약속_생성_요청 = new AppointmentCreateRequest(수현_약속_제목, 하루_뒤_시간, 수현_약속_위도, 수현_약속_경도, 3L, 스터디_카테고리_제목, 스터디_카테고리_컬러_문자);
     public static final AppointmentParticipateRequest 수현_약속_참여_요청 = new AppointmentParticipateRequest(수현_약속_초대코드.getCode());
+    public static final AppointmentResponse 수현_약속_응답 = new AppointmentResponse(1L, 1L, 수현_약속_제목, 하루_뒤_시간, 수현_약속_위도, 수현_약속_경도, 수현_약속_초대코드.getCode(), 3L);
 
     public static Appointment 수현_약속() {
         Appointment appointment = new Appointment(수현_약속_제목, 수현_약속_초대코드, 하루_뒤_시간, 3L);
@@ -34,6 +36,7 @@ public class AppointmentFixtures {
     public static final BigDecimal 민성_약속_위도 = BigDecimal.valueOf(126.99597295767953);
     public static final BigDecimal 민성_약속_경도 = BigDecimal.valueOf(37.5280674292228);
     public static final AppointmentCreateRequest 민성_약속_요청 = new AppointmentCreateRequest(민성_약속_제목, 하루_뒤_시간, 민성_약속_위도, 민성_약속_경도, 3L, 스터디_카테고리_제목, 스터디_카테고리_컬러_문자);
+    public static final AppointmentResponse 민성_약속_응답 = new AppointmentResponse(2L, 2L, 민성_약속_제목, 하루_뒤_시간, 민성_약속_위도, 민성_약속_경도, 민성_약속_초대코드.getCode(), 3L);
 
     public static Appointment 민성_약속() {
         Appointment appointment = new Appointment(민성_약속_제목, 민성_약속_초대코드, 하루_뒤_시간, 3L);
